Added AppBody component Simplified css across components
