+++ src/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <body>
         <tn-header></tn-header>
-        <tn-tuner></tn-tuner>
+        <tn-app></tn-app>
 
         <script src="main.js"></script>
         <link rel="stylesheet" href="main.css">
